+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### [13.34.0](https://kaos.sh/ek/13.34.0)
+
+- **`[i18n]`** Type `String` has been renamed to `Text`
+- **`[i18n]`** Added method `Text.Format`
+
 ### [13.33.1](https://kaos.sh/ek/13.33.1)
 
 - **`[fmtc]`** Code refactoring